requirements.txt
elasticsearch
PyYAML
scikit-learn
matplotlib
pandas
seaborn
